requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.2
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contextlib2==21.6.0
cron_descriptor==2.0.6
cryptography==44.0.2
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.7
django-allauth==65.6.0
django-celery-beat==2.8.1
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-htmx==1.23.0
django-import-export==4.3.9
django-jazzmin==3.0.1
django-ninja==1.4.1
django-ninja-extra==0.30.0
django-ninja-jwt==5.3.7
django-redis==6.0.0
django-silk==5.4.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.10
et_xmlfile==2.0.0
factory_boy==3.3.3
Faker==37.4.0
git-filter-repo==2.47.0
gprof2dot==2025.4.14
gunicorn==23.0.0
hiredis==3.2.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
injector==0.22.0
isort==6.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
narwhals==1.48.0
numpy==2.3.1
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
pillow==11.1.0
plotly==6.2.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.2
PyJWT==2.9.0
pytest==8.4.1
pytest-django==4.11.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.27.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
stripe==12.4.0
tablib==3.8.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.9.0
